==> scripts/2.Deeplearning_Celltype_and_PMD/7.2_Image_To_CellSighter_Prediction_celltype.ipynb <==
import sys
sys.path.append(".")
from torch.utils.tensorboard import SummaryWriter
import os
import torch
import argparse
import numpy as np
import json
import os
import pandas as pd
import numpy as np
from tifffile import imread
import random
import cv2
import tifffile
from skimage import exposure
import matplotlib
import matplotlib.pyplot as plt
import numpy as np
import re
from skimage import data, img_as_float
from skimage import exposure
import qnorm
import random

output_dir = "../output/CellSighter/celltype/"
cell_marker_file = "../annotation/cell_markers.csv"
tiff_dir = "../output/input/"
segmentation_dir  = "../output/segmentation/"
cell_table_dir  = "../output/cell_table/"

cell_type = pd.read_csv(os.path.join(cell_marker_file))

cell_types_unique = cell_type.cell_type[np.sort(np.unique(cell_type.cell_type, return_index=True)[1])]
cell_types_unique = np.flip(cell_types_unique)
marker_unique = pd.read_csv(os.path.join(output_dir, "cell_classification", "channels.txt"), header=None)
marker_unique = marker_unique[0].values

mat = np.zeros(shape = (2000,2000, len(os.listdir(tiff_dir)[0:28])))

# Calculate mat for quantile normalization
mat_dict = dict()
for marker in marker_unique:
    print(marker)
    mat = np.zeros(shape = (2000,2000, len(os.listdir(tiff_dir)[0:28])))
    i=0
    for sample in os.listdir(tiff_dir)[0:28]:
        print(sample)
        img = imread(os.path.join(os.path.join(tiff_dir, sample, marker + ".tiff")))
        mat[...,i] = img[0:2000,0:2000]
        i = i + 1
    random.shuffle(mat)
    mat = mat.reshape(28*2000*2000)
    mat_dict[marker] = mat

for sample in ["ROI-23", "ROI-27", "ROI-59", "ROI-60", "ROI-61"]:
    print(sample)
    if not os.path.exists(output_dir):
        os.makedirs(output_dir)
    base_dir = os.path.join(output_dir, "Predictions")
    if not os.path.exists(base_dir):
        os.makedirs(base_dir)
    cellType_dir = os.path.join(base_dir, "CellTypes")
    if not os.path.exists(cellType_dir):
        os.makedirs(cellType_dir)
    cells_dir = os.path.join(cellType_dir, "cells")
    if not os.path.exists(cells_dir):
        os.makedirs(cells_dir)
    cells2labels_dir = os.path.join(cellType_dir, "cells2labels")
    if not os.path.exists(cells2labels_dir):
        os.makedirs(cells2labels_dir)
    data_dir = os.path.join(cellType_dir, "data")
    if not os.path.exists(data_dir):
        os.makedirs(data_dir)
    images_dir =  os.path.join(data_dir, "images")
    if not os.path.exists(images_dir):
        os.makedirs(images_dir)    

    whole_cell = imread(os.path.join(segmentation_dir, sample + "_whole_cell.tiff"))
    np.savez(os.path.join(cells_dir, sample +".npz"), data = whole_cell)
    
    cells = pd.read_csv(os.path.join(cell_table_dir, sample + "_cell_table_size_normalized.csv.gz"))
    labels = np.zeros(int(max(cells.label)) + 1)
    for i in range(len(labels)):
        labels[i] = -1
    np.savez(os.path.join(cells2labels_dir,  sample + ".npz"), data = labels)
    
    # Copy full markers images

    all_markers = []
    # Load the markers 
    for marker in marker_unique:
        print(marker)
        
        
        if os.path.isfile(os.path.join(tiff_dir, sample, marker + ".tiff")):
            
            num = int(re.sub(".*-", "", sample))
            if num > 28:
                print("Scaling " + marker + " !")
                img = imread(os.path.join(os.path.join(tiff_dir, sample, marker + ".tiff")))
                values = img.squeeze().reshape(img.shape[0]*img.shape[1])
                mat_quant = mat_dict[marker][0:len(values)]
                values_df = pd.DataFrame(values)
                mat_quant_df = pd.DataFrame(mat_quant)
                norm_values_df = qnorm.quantile_normalize(values_df, target = mat_quant, ncpus=10)
                norm_values = norm_values_df[0].values
                norm_values = norm_values.astype(int)
                marker_image = norm_values.reshape(img.shape[0],img.shape[1])
                marker_image = np.int32(marker_image)                      
            else:
                # Load segmentation 
                marker_image = imread(os.path.join(tiff_dir, sample, marker + ".tiff"))


            # Save in the "cells" folder
            all_markers.append(marker_image)
        else:
            # Load segmentation 
            marker_image = imread(os.path.join(tiff_dir, sample, "DAPI" + ".tiff"))
            marker_image[marker_image>0]=0
            # Crop to the right corner region
            # marker_cropped = marker_image[y1:y2, x1:x2]

            # Save in the "cells" folder
            all_markers.append(marker_image)


    # Combine
    all_markers = np.array(all_markers)
    all_markers = np.transpose(all_markers, (1, 2, 0))

    # Save as npz in the "data" folder

    np.savez(os.path.join(images_dir,  sample + ".npz"),  data = all_markers)
